examples/crewai/requirements.txt
crewai
httpx
web3
python-dotenv
pydantic
langchain
